refactor: move inline swiftui view into separate file
